server/requirements.txt
Flask == 3.1.0
Flask-Cors == 5.0.0
Flask-Limiter == 3.10.1
flask-simple-captcha == 5.5.5
gunicorn == 23.0.0
redis == 5.2.1
ulid-py == 1.1.0